+++ src/Compilers/Core/Portable/System/Reflection/Metadata/Ecma335/MetadataBuilder.Heaps.cs
@@ -228,7 +228,7 @@ namespace Roslyn.Reflection.Metadata.Ecma335
                 }
 
                 _userStrings.Add(str, index);
-                _userStringWriter.WriteCompressedInteger((uint)str.Length * 2 + 1);
+                _userStringWriter.WriteCompressedInteger(str.Length * 2 + 1);
 
                 _userStringWriter.WriteUTF16(str);
 
@@ -401,7 +401,7 @@ namespace Roslyn.Reflection.Metadata.Ecma335
                 var blob = entry.Key;
 
                 writer.Offset = heapOffset;
-                writer.WriteCompressedInteger((uint)blob.Length);
+                writer.WriteCompressedInteger(blob.Length);
                 writer.WriteBytes(blob);
             }
